feat(streak): week helpers and reading streak UI updates
Refactor streak button/popup and add readingStreakWeekHelpers. Adjust
AgentsHighlightsSection for highlights layout.

Made-with: Cursor
