+++ packages/shared/src/components/streak/popup/ReadingStreakPopup.tsx
@@ -1,11 +1,10 @@
 import type { ReactElement } from 'react';
 import React, { useEffect, useMemo } from 'react';
-import { addDays, subDays } from 'date-fns';
 import { useQuery } from '@tanstack/react-query';
 import classNames from 'classnames';
 import { useRouter } from 'next/router';
 import { StreakSection } from './StreakSection';
-import { DayStreak, Streak } from './DayStreak';
+import { DayStreak } from './DayStreak';
 import { generateQueryKey, RequestKey, StaleTime } from '../../../lib/query';
 import type { ReadingDay, UserStreak } from '../../../graphql/users';
 import { getReadingStreak30Days } from '../../../graphql/users';
@@ -14,12 +13,13 @@ import { useActions, useViewSize, ViewSize } from '../../../hooks';
 import { ActionType } from '../../../graphql/actions';
 import { Button, ButtonSize, ButtonVariant } from '../../buttons/Button';
 import { SettingsIcon, VIcon, WarningIcon } from '../../icons';
-import { isWeekend, DayOfWeek } from '../../../lib/date';
+import { DayOfWeek } from '../../../lib/date';
 import {
   DEFAULT_TIMEZONE,
   getTimezoneOffsetLabel,
   isSameDayInTimezone,
 } from '../../../lib/timezones';
+import { getStreak, getStreakDays } from '../readingStreakWeekHelpers';
 import { timezoneSettingsUrl, webappUrl } from '../../../lib/constants';
 import { useStreakTimezoneOk } from '../../../hooks/streaks/useStreakTimezoneOk';
 import { usePrompt } from '../../../hooks/usePrompt';
@@ -47,60 +47,6 @@ import { NotificationSvg } from '../../notifications/NotificationSvg';
 import { usePushNotificationContext } from '../../../contexts/PushNotificationContext';
 import { IconSize } from '../../Icon';
 import { Tooltip } from '../../tooltip/Tooltip';
-
-const getStreak = ({
-  value,
-  today,
-  history,
-  startOfWeek = DayOfWeek.Monday,
-  timezone,
-}: {
-  value: Date;
-  today: Date;
-  history?: ReadingDay[];
-  startOfWeek?: number;
-  timezone?: string;
-}): Streak => {
-  const isFreezeDay = isWeekend(value, startOfWeek, timezone);
-  const isToday = isSameDayInTimezone(value, today, timezone);
-  const isFuture = value > today;
-  const isCompleted =
-    !isFuture &&
-    history?.some(({ date: historyDate, reads }) => {
-      const dateToCompare = new Date(historyDate);
-      const sameDate = isSameDayInTimezone(dateToCompare, value, timezone);
-
-      return sameDate && reads > 0;
-    });
-
-  if (isCompleted) {
-    return Streak.Completed;
-  }
-
-  if (isFreezeDay) {
-    return Streak.Freeze;
-  }
-
-  if (isToday) {
-    return Streak.Pending;
-  }
-
-  return Streak.Upcoming;
-};
-
-const getStreakDays = (today: Date) => {
-  return [
-    subDays(today, 4),
-    subDays(today, 3),
-    subDays(today, 2),
-    subDays(today, 1),
-    today,
-    addDays(today, 1),
-    addDays(today, 2),
-    addDays(today, 3),
-    addDays(today, 4),
-  ]; // these dates will then be compared to the user's post views
-};
 
 interface ReadingStreakPopupProps {
   streak: UserStreak;
